Renamed the usersService to UsersApiService for better clarity
